+++ rickled/cli.py
@@ -1,21 +1,173 @@
+# ASCCI Text found in this file was generated on
+# https://patorjk.com/software/taag/
+# Large heading using "Shaded Blocky" font
+# Sub headings using "Small" font
+
 import rickled.__version__ as ver
 import argparse
-
 from rickled.tools import bcolors
+from rickled.tools import Schema
+from rickled.tools import Converter
+from rickled.net import serve_rickle_http, serve_rickle_https
+from rickled import Rickle
+import re
+import json
+import yaml
+import ast
+
+
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
 
 def conv(args):
-    from rickled.tools import Converter
     Converter(input_files=args.i,
               output_files=args.o,
               input_directories=args.d,
               default_output_type=args.t,
               silent=args.s).do_convert()
 
+def obj_get(args):
+    import json
+    import yaml
+    if args:
+        r = Rickle(args.i, load_lambda=args.l)
+        v = r.get(args.key)
+        if isinstance(v, Rickle):
+            if args.t.lower() == 'json':
+                print(v.to_json_string())
+            else:
+                print(v.to_yaml_string())
+        elif isinstance(v, dict):
+            if args.t.lower() == 'json':
+                print(json.dumps(v))
+            else:
+                print(yaml.safe_dump(v))
+        else:
+            print(v)
+
+
+
+def obj_set(args):
+    if args:
+        r = Rickle(args.i)
+        r.set(args.key, args.value)
+        if args.p:
+            if args.t.lower() == 'json':
+                print(r.to_json_string())
+            else:
+                print(r.to_yaml_string())
+        if args.o:
+            if args.t.lower() == 'json':
+                r.to_json_file(args.o)
+            else:
+                r.to_yaml_file(args.o)
+
+
+def obj_del(args):
+    if args:
+        r = Rickle(args.i)
+        r.remove(args.key)
+        if args.p:
+            if args.t.lower() == 'json':
+                print(r.to_json_string())
+            else:
+                print(r.to_yaml_string())
+        if args.o:
+            if args.t.lower() == 'json':
+                r.to_json_file(args.o)
+            else:
+                r.to_yaml_file(args.o)
+
+def obj_type(args):
+    if args:
+        r = Rickle(args.i, load_lambda=args.l)
+        v = r.get(args.key)
+        print(type(v))
+
+def obj_search(args):
+    if args:
+        r = Rickle(args.i, load_lambda=args.l)
+        paths = r.search_path(args.key)
+        for p in paths:
+            print(p)
+
+def obj_func(args):
+    def guess_parse(param_value):
+        try:
+            return ast.literal_eval(param_value)
+        except Exception as exc:
+            raise TypeError(f"Could not guess the parameter type, {exc}")
+    def parse_type(type_name, param_value):
+        type_name = type_name.strip().lower()
+        if type_name == 'str':
+            return str(param_value)
+        elif type_name == 'int':
+            try:
+                return int(param_value)
+            except:
+                raise TypeError(f"Value '{param_value}' does not match type int")
+        elif type_name == 'float':
+            try:
+                return float(param_value)
+            except:
+                raise TypeError(f"Value '{param_value}' does not match type int")
+        elif type_name == 'bool':
+            if param_value.strip().lower() == 'true':
+                return True
+            if param_value.strip().lower() == 'false':
+                return False
+            raise TypeError(f"Value '{param_value}' does not match type bool")
+        elif type_name == 'list' or type_name == 'dict':
+            try:
+                return ast.literal_eval(param_value)
+            except Exception as exc:
+                raise TypeError(f"Could not guess the parameter type, {exc}")
+        else:
+            raise ValueError('Could not interpret type, only str, int, float, bool, list, dict accepted.')
+
+
+    re_pat = re.compile(r"(.+?)=(.+)")
+    if args:
+        r = Rickle(args.i, load_lambda=args.l)
+
+        params = dict()
+        if args.params:
+            for p in args.params:
+                m = re_pat.match(p)
+                param_name = m.group(1)
+                param_value = m.group(2)
+                if ':' in param_name:
+                    param_name, ptype = param_name.split(':')
+                    param_value = parse_type(ptype, param_value)
+                elif args.x:
+                    param_value = guess_parse(param_value)
+
+                params[param_name] = param_value
+        v = r(args.key, **params)
+
+        if not v is None:
+            if isinstance(v, Rickle):
+                if args.t.lower() == 'json':
+                    print(v.to_json_string())
+                else:
+                    print(v.to_yaml_string())
+            elif isinstance(v, dict):
+                if args.t.lower() == 'json':
+                    print(json.dumps(v))
+                else:
+                    print(yaml.safe_dump(v))
+            else:
+                print(v)
+
+
 def serve(args):
-    from rickled.net import serve_rickle_http, serve_rickle_https
-    from rickled import Rickle
-
-
     rick = Rickle(args.f)
 
     if args.b:
@@ -38,7 +190,6 @@ def serve(args):
                           )
 
 def check(args):
-    from rickled.tools import Schema
     Schema(input_files=args.i,
            input_directories=args.d,
            schema=args.c,
@@ -46,7 +197,6 @@ def check(args):
            silent=args.s).do_validation()
 
 def gen(args):
-    from rickled.tools import Schema
     Schema(input_files=args.i,
            input_directories=args.d,
            silent=args.s).do_generation()
@@ -72,6 +222,12 @@ def main():
     subparsers = parser.add_subparsers()
 
     #################### CONV #####################
+    # ░░      ░░░      ░░   ░░░  ░  ░░░░  ░
+    # ▒  ▒▒▒▒  ▒  ▒▒▒▒  ▒    ▒▒  ▒  ▒▒▒▒  ▒
+    # ▓  ▓▓▓▓▓▓▓  ▓▓▓▓  ▓  ▓  ▓  ▓▓  ▓▓  ▓▓
+    # █  ████  █  ████  █  ██    ███    ███
+    # ██      ███      ██  ███   ████  ████
+
 
     parser_conv = subparsers.add_parser('conv',
                                         help=f'Tool for {bcolors.OKBLUE}converting{bcolors.ENDC} files to or from YAML',
@@ -94,7 +250,171 @@ def main():
 
     parser_conv.set_defaults(func=conv)
 
+    #################### OBJ #####################
+    # ░░      ░░       ░░        ░
+    # ▒  ▒▒▒▒  ▒  ▒▒▒▒  ▒▒▒▒▒▒▒  ▒
+    # ▓  ▓▓▓▓  ▓       ▓▓▓▓▓▓▓▓  ▓
+    # █  ████  █  ████  █  ████  █
+    # ██      ██       ███      ██
+
+
+    parser_obj = subparsers.add_parser('obj',
+                                        help=f'Tool for {bcolors.OKBLUE}accessing/manipulating{bcolors.ENDC} YAML files',
+                                        description=f"""
+    ---
+    {bcolors.HEADER}Tool for accessing/manipulating YAML files{bcolors.ENDC}.
+        """,
+                                        epilog="-----------------------------------------------------"
+                                        )
+
+    parser_obj.add_argument('-i', type=str, help=f"{bcolors.OKBLUE}input file{bcolors.ENDC} to read/modify",
+                            metavar='input', required=True)
+    parser_obj.add_argument('-o', type=str, help=f"{bcolors.OKBLUE}output file{bcolors.ENDC} to save modified",
+                            metavar='output', required=False)
+    parser_obj.add_argument('-t', type=str,
+                             help=f"output {bcolors.OKBLUE}type{bcolors.ENDC} (JSON, YAML)",
+                             default='yaml', metavar='type')
+    parser_obj.add_argument('-p', action='store_true', help=f"{bcolors.OKBLUE}print{bcolors.ENDC} output",
+                            default=False)
+    parser_obj.add_argument('-l', action='store_true', help=f"load {bcolors.OKBLUE}lambda{bcolors.ENDC} types",
+                            default=False)
+
+    subparsers_obj = parser_obj.add_subparsers()
+
+    #################### OBJ - GET #####################
+    #   ___ ___ _____
+    #  / __| __|_   _|
+    # | (_ | _|  | |
+    #  \___|___| |_|
+
+    get_obj_parser = subparsers_obj.add_parser('get',
+                       help=f'Tool for {bcolors.OKBLUE}getting{bcolors.ENDC} values from YAML files',
+                       description=f"""
+        ---
+        {bcolors.HEADER}Tool for getting values from YAML files{bcolors.ENDC}.
+            """,
+                                          epilog="-----------------------------------------------------"
+                                          )
+
+    get_obj_parser.add_argument('key', type=str, help=f"{bcolors.OKBLUE}Key{bcolors.ENDC} to get value",
+                            metavar='key')
+
+    get_obj_parser.set_defaults(func=obj_get)
+
+    #################### OBJ - SET #####################
+    #  ___ ___ _____
+    # / __| __|_   _|
+    # \__ \ _|  | |
+    # |___/___| |_|
+
+    set_obj_parser = subparsers_obj.add_parser('set',
+                                               help=f'Tool for {bcolors.OKBLUE}setting{bcolors.ENDC} values in YAML files',
+                                               description=f"""
+            ---
+            {bcolors.HEADER}Tool for setting values in YAML files{bcolors.ENDC}.
+                """,
+                                               epilog="-----------------------------------------------------"
+                                               )
+
+    set_obj_parser.add_argument('key', type=str, help=f"{bcolors.OKBLUE}Key{bcolors.ENDC} to set value",
+                                metavar='key')
+    set_obj_parser.add_argument('value', type=str, help=f"{bcolors.OKBLUE}Value{bcolors.ENDC} to set",
+                                metavar='value')
+
+    set_obj_parser.set_defaults(func=obj_set)
+
+    #################### OBJ - DEL #####################
+    #  ___  ___ _
+    # |   \| __| |
+    # | |) | _|| |__
+    # |___/|___|____|
+
+    del_obj_parser = subparsers_obj.add_parser('del',
+                                               help=f'Tool for {bcolors.OKBLUE}deleting{bcolors.ENDC} keys in YAML files',
+                                               description=f"""
+                ---
+                {bcolors.HEADER}Tool for deleting keys in YAML files{bcolors.ENDC}.
+                    """,
+                                               epilog="-----------------------------------------------------"
+                                               )
+
+    del_obj_parser.add_argument('key', type=str, help=f"{bcolors.OKBLUE}Key{bcolors.ENDC} to delete",
+                                metavar='key')
+
+    del_obj_parser.set_defaults(func=obj_del)
+
+    #################### OBJ - TYPE #####################
+    #  _______   _____ ___
+    # |_   _\ \ / / _ \ __|
+    #   | |  \ V /|  _/ _|
+    #   |_|   |_| |_| |___|
+
+    type_obj_parser = subparsers_obj.add_parser('type',
+                                               help=f'Tool for {bcolors.OKBLUE}checking type{bcolors.ENDC} of keys in YAML files',
+                                               description=f"""
+                    ---
+                    {bcolors.HEADER}Tool for checking type of keys in YAML files{bcolors.ENDC}.
+                        """,
+                                               epilog="-----------------------------------------------------"
+                                               )
+
+    type_obj_parser.add_argument('key', type=str, help=f"{bcolors.OKBLUE}Key{bcolors.ENDC} to check",
+                                metavar='key')
+
+    type_obj_parser.set_defaults(func=obj_type)
+
+    #################### OBJ - SEARCH #####################
+    #  ___ ___   _   ___  ___ _  _
+    # / __| __| /_\ | _ \/ __| || |
+    # \__ \ _| / _ \|   / (__| __ |
+    # |___/___/_/ \_\_|_\\___|_||_|
+
+    search_obj_parser = subparsers_obj.add_parser('search',
+                                               help=f'Tool for {bcolors.OKBLUE}searching{bcolors.ENDC} keys in YAML files',
+                                               description=f"""
+                    ---
+                    {bcolors.HEADER}Tool for searching keys in YAML files{bcolors.ENDC}.
+                        """,
+                                               epilog="-----------------------------------------------------"
+                                               )
+
+    search_obj_parser.add_argument('key', type=str, help=f"{bcolors.OKBLUE}Key{bcolors.ENDC} to search",
+                                metavar='key')
+
+    search_obj_parser.set_defaults(func=obj_search)
+
+    #################### OBJ - FUNC #####################
+    #  ___ _   _ _  _  ___
+    # | __| | | | \| |/ __|
+    # | _|| |_| | .` | (__
+    # |_|  \___/|_|\_|\___|
+
+    func_obj_parser = subparsers_obj.add_parser('func',
+                                               help=f'Tool for {bcolors.OKBLUE}executing function{bcolors.ENDC} defined in YAML files',
+                                               description=f"""
+                    ---
+                    {bcolors.HEADER}Tool for executing function defined in YAML files{bcolors.ENDC}.
+                        """,
+                                               epilog="-----------------------------------------------------"
+                                               )
+
+    func_obj_parser.add_argument('-x', action='store_true', help=f"{bcolors.OKBLUE}infer parameter{bcolors.ENDC} types",
+                            default=False)
+    func_obj_parser.add_argument('key', type=str, help=f"{bcolors.OKBLUE}Key{bcolors.ENDC} (name) of function",
+                                metavar='key')
+    func_obj_parser.add_argument('params', type=str, help=f"{bcolors.OKBLUE}Params{bcolors.ENDC} for function",
+                                 metavar='params', nargs='+')
+
+    func_obj_parser.set_defaults(func=obj_func)
+
+
     #################### SERVE #####################
+    # ░░      ░░        ░       ░░  ░░░░  ░        ░
+    # ▒  ▒▒▒▒▒▒▒  ▒▒▒▒▒▒▒  ▒▒▒▒  ▒  ▒▒▒▒  ▒  ▒▒▒▒▒▒▒
+    # ▓▓      ▓▓      ▓▓▓       ▓▓▓  ▓▓  ▓▓      ▓▓▓
+    # ███████  █  ███████  ███  ████    ███  ███████
+    # ██      ██        █  ████  ████  ████        █
+
 
     parser_serve = subparsers.add_parser('serve',
                                         help=f'Tool for serving YAML via {bcolors.OKBLUE}http(s){bcolors.ENDC}',
@@ -123,6 +443,12 @@ def main():
     parser_serve.set_defaults(func=serve)
 
     #################### SCHEMA #####################
+    # ░░      ░░░      ░░  ░░░░  ░        ░  ░░░░  ░░      ░░
+    # ▒  ▒▒▒▒▒▒▒  ▒▒▒▒  ▒  ▒▒▒▒  ▒  ▒▒▒▒▒▒▒   ▒▒   ▒  ▒▒▒▒  ▒
+    # ▓▓      ▓▓  ▓▓▓▓▓▓▓        ▓      ▓▓▓        ▓  ▓▓▓▓  ▓
+    # ███████  █  ████  █  ████  █  ███████  █  █  █        █
+    # ██      ███      ██  ████  █        █  ████  █  ████  █
+
 
     parser_schema = subparsers.add_parser('schema',
                                         help=f'Tool for generating and checking {bcolors.OKBLUE}schemas{bcolors.ENDC} of YAML files',
@@ -136,6 +462,11 @@ def main():
     schema_subparsers = parser_schema.add_subparsers()
 
     #################### SCHEMA - CHECK ############
+    #   ___ _  _ ___ ___ _  __
+    #  / __| || | __/ __| |/ /
+    # | (__| __ | _| (__| ' <
+    #  \___|_||_|___\___|_|\_\
+
     parser_schema_check = schema_subparsers.add_parser('check',
                                           help=f'Tool for {bcolors.OKBLUE}checking{bcolors.ENDC} schemas of YAML files',
                                           description=f"""
@@ -162,9 +493,12 @@ def main():
 
     parser_schema_check.set_defaults(func=check)
 
-    #################################################
-
     #################### SCHEMA - GEN ############
+    #   ___ ___ _  _
+    #  / __| __| \| |
+    # | (_ | _|| .` |
+    #  \___|___|_|\_|
+
     parser_schema_gen = schema_subparsers.add_parser('gen',
                                                        help=f'Tool for {bcolors.OKBLUE}generating{bcolors.ENDC} schemas of YAML files',
                                                        description=f"""